+++ python/ml4ir/applications/classification/pipeline.py
@@ -1,12 +1,8 @@
-import os
 import sys
 import ast
 from argparse import Namespace
-import pandas as pd
 from tensorflow.keras.metrics import Metric
 from tensorflow.keras.optimizers import Optimizer
-from tensorflow import data
-from ml4ir.base.model.relevance_model import RelevanceModelConstants
 from ml4ir.applications.classification.config.parse_args import get_args
 from ml4ir.applications.classification.model.losses import categorical_cross_entropy
 from ml4ir.applications.classification.model.metrics import metrics_factory
@@ -18,8 +14,8 @@ from ml4ir.base.model.relevance_model import RelevanceModel
 from ml4ir.base.model.scoring.scoring_model import ScorerBase, RelevanceScorer
 from ml4ir.base.model.scoring.interaction_model import InteractionModel, UnivariateInteractionModel
 from ml4ir.base.pipeline import RelevancePipeline
-
-from typing import Union, List, Type, Optional
+from ml4ir.applications.classification.model.classification_model import ClassificationModel
+from typing import Union, List, Type
 
 
 class ClassificationPipeline(RelevancePipeline):
@@ -98,7 +94,7 @@ class ClassificationPipeline(RelevancePipeline):
         )
 
         # Combine the above to define a RelevanceModel
-        relevance_model: RelevanceModel = RelevanceModel(
+        relevance_model: RelevanceModel = ClassificationModel(
             feature_config=self.feature_config,
             scorer=scorer,
             metrics=metrics,
@@ -166,102 +162,6 @@ class ClassificationPipeline(RelevancePipeline):
 
         return relevance_dataset
 
-    def evaluate(
-        self,
-        test_dataset: data.TFRecordDataset,
-        inference_signature: str = None,
-        additional_features: dict = {},
-        group_metrics_min_queries: int = 50,
-        logs_dir: Optional[str] = None,
-        logging_frequency: int = 25,
-    ):
-        """
-        Evaluate the Classification Model
-
-        Parameters
-        ----------
-        test_dataset: an instance of tf.data.dataset
-        inference_signature : str, optional
-            If using a SavedModel for prediction, specify the inference signature to be used for computing scores
-        additional_features : dict, optional
-            Dictionary containing new feature name and function definition to
-            compute them. Use this to compute additional features from the scores.
-            For example, converting ranking scores for each document into ranks for
-            the query
-        group_metrics_min_queries : int, optional
-            Minimum count threshold per group to be considered for computing
-            groupwise metrics
-        logs_dir : str, optional
-            Path to directory to save logs
-        logging_frequency : int
-            Value representing how often(in batches) to log status
-
-        Returns
-        -------
-        df_overall_metrics : `pd.DataFrame` object
-            `pd.DataFrame` containing overall metrics
-        df_groupwise_metrics : `pd.DataFrame` object
-            `pd.DataFrame` containing groupwise metrics if
-            group_metric_keys are defined in the FeatureConfig
-        metrics_dict : dict
-            metrics as a dictionary of metric names mapping to values
-
-        Notes
-        -----
-        You can directly do a `model.evaluate()` only if the keras model is compiled
-        """
-        relevance_model = self.get_relevance_model()
-        if not relevance_model.is_compiled:
-            return NotImplementedError
-
-        group_metrics_keys = self.feature_config.get_group_metrics_keys()
-
-        metrics_dict = relevance_model.model.evaluate(test_dataset)
-        metrics_dict = dict(zip(relevance_model.model.metrics_names, metrics_dict))
-
-        predictions = relevance_model.predict(test_dataset,
-                                              inference_signature=inference_signature,
-                                              additional_features=additional_features,
-                                              logs_dir=None,  # Return pd.DataFrame of predictions
-                                              logging_frequency=logging_frequency)
-        # Need to convert predictions from tuples of tf.Tensor to lists of int/float
-        # so that we are compatible with tf.keras.metrics
-        label_name = self.feature_config.get_label()['name']
-        output_name = relevance_model.output_name
-        predictions[label_name] = predictions[label_name].apply(lambda l: [item.numpy() for item in l])
-        predictions[output_name] = predictions[output_name].apply(lambda l: [item.numpy() for item in l])
-
-        global_metrics = []  # group_name, metric, value
-        grouped_metrics = []
-        for metric in relevance_model.model.metrics:
-            metric.reset_states()
-            metric.update_state(predictions[label_name].values.tolist(), predictions[output_name].values.tolist())
-            global_metrics.append({"metric": metric.name,
-                                   "value": metric.result().numpy()})
-            for group_ in group_metrics_keys:
-                for name, group in predictions.groupby(group_['name']):
-                    if group.shape[0] >= group_metrics_min_queries:
-                        metric.reset_states()
-                        metric.update_state(group[label_name].values.tolist(),
-                                            group[output_name].values.tolist())
-                        grouped_metrics.append({"group_name": group_['name'],
-                                                "group_key": name,
-                                                "metric": metric.name,
-                                                "value": metric.result().numpy(),
-                                                "size": group.shape[0]})
-        global_metrics = pd.DataFrame(global_metrics)
-        grouped_metrics = pd.DataFrame(grouped_metrics)
-        if logs_dir:
-            self.file_io.write_df(
-                grouped_metrics,
-                outfile=os.path.join(logs_dir, RelevanceModelConstants.GROUP_METRICS_CSV_FILE),
-                )
-            self.file_io.write_df(
-                global_metrics,
-                outfile=os.path.join(logs_dir, RelevanceModelConstants.METRICS_CSV_FILE)
-            )
-            self.logger.info(f"Evaluation Results written at: {logs_dir}")
-        return global_metrics, grouped_metrics, metrics_dict
 
 
 def main(argv):
